+++ studio/schemas/post.js
@@ -9,7 +9,8 @@ export default {
     {
       title: 'Title',
       name: 'title',
-      type: 'string'
+      type: 'string',
+      validation: (Rule) => Rule.required()
     },
     {
       title: 'Slug',
@@ -18,7 +19,8 @@ export default {
       options: {
         source: 'title',
         maxLength: 96
-      }
+      },
+      validation: (Rule) => Rule.required()
     },
     {
       title: 'Main Image',